Add some postprocessing tools
