+++ test/basic_wing.jl
@@ -4,15 +4,18 @@ Pkg.activate(".")
 using Revise
 using AeroMDAO
 using LinearAlgebra
+using GLMakie
+Makie.inline!(false)
+# using Makie.GeometryBasics
 
 ## Geometry
 wing = WingSection(root_foil  = naca4(0,0,1,2),
                    tip_foil   = naca4(0,0,1,2),
                    root_chord = 1.0,
-                   taper      = 1.0,
+                   taper      = 0.4,
                    span       = 5.0,
-                   dihedral   = 0.0,
-                   sweep      = 0.0)
+                   dihedral   = 5.0,
+                   sweep      = 25.0)
 
 # Meshing
 wing_mesh = WingMesh(wing, [20], 20; span_spacing = Uniform())
@@ -29,49 +32,15 @@ fs = Freestream( α, β, zeros(3))
 
 ##
 @time prob = solve_system_neumann(surf_pans, Umag, fs, 50)
-# φs = permutedims(
-#     reshape(
-#         prob.singularities[1:length(prob.surface_panels)], 
-#         size(prob.surface_panels,2), size(prob.surface_panels,1))
-# );
-
-##
 @time vs = surface_velocities(prob);
-@time CL, CD = surface_coefficients(prob, wing);
+@time CL, CD, CP = surface_coefficients(prob, wing);
 println("Σᵢ CLᵢ: $CL")
 println("Σᵢ CDᵢ: $CD")
 
 ## Plotting
-using Plots
-plotly()
-plt_surfs = plot_panels(surf_pans)
+f = Figure()
+ax = LScene(f[1, 1])
+pts, cnt, C = plot_scalar_field(surf_pans, CP)
+poly!(pts, cnt; color=C[:], strokewidth=0.1)
+current_figure()
 
-xl=-1;xh=4;yl=-3;yh=3;zl=-1;zh=1;
-p = Plots.plot(; xlabel="x",ylabel="y",zlabel="z", grid=:true, xlim = (xl, xh), ylim = (yl, yh), zlim=(zl, zh))
-
-for pan in plt_surfs
-    p = Plots.plot!(pan, color = :grey)
-end
-
-plt_surfs = plot_panels(prob.wake_panels)
-for pan in plt_surfs
-    p = Plots.plot!(pan, color = :blue)
-end
-
-plot!(
-    getindex.(collocation_point.(surf_pans), 1),
-    getindex.(collocation_point.(surf_pans), 2),
-    getindex.(collocation_point.(surf_pans), 3),
-    st = :surface
-)
-
-plot!(
-    getindex.(surf_pts, 1),
-    getindex.(surf_pts, 2),
-    getindex.(surf_pts, 3),
-    st = :surface)
-
-plot!(p)
-# Plots.scatter!(Tuple.(surf_pts)[:], markersize = 0.1)
-
-##
